+++ src/graph/geograph.py
@@ -12,7 +12,6 @@ from graph.base import Digraph
 from graph.astar import a_star
 from utils.serialization import save_checkpoint, load_checkpoint
 
-#%%
 class GeoDigraph(Digraph):
     def __init__(self, df_edges:GeoDataFrame=None, df_nodes:GeoDataFrame=None, *args, **kwargs):
         self.df_edges = df_edges
@@ -24,7 +23,7 @@ class GeoDigraph(Digraph):
             super().__init__(df_edges[['src', 'dst', 'dist']].values, df_nodes.to_dict(orient='index'), *args, **kwargs)
         
 
-    def search(self, src, dst, max_steps=2000, max_dist=10000, geometry=True):
+    def search(self, src, dst, max_steps=2000, max_dist=10000, geom=True):
         # TODO: (src, dst) 最短路径 geometry
         route = a_star(graph=self.graph,
                        nodes=self.nodes,
@@ -35,8 +34,16 @@ class GeoDigraph(Digraph):
                        max_steps=max_steps,
                        max_dist=max_dist
         )
-        # if geometry:
-        #     route['geometry'] = self.transform_node_seq_to_polyline(route['path'])
+        
+        if 'eid_list' not in route:
+            route['eid_list'] = self.transform_node_seq_to_edge_seq(route['path'])
+        
+        if geom and 'geometry' not in route:
+            lst = route['eid_list']
+            if lst is None:
+                route['geometry'] = None
+            else:
+                route['geometry'] = self.transform_edge_seq_to_polyline(lst)
         
         return route
 
@@ -50,10 +57,10 @@ class GeoDigraph(Digraph):
 
 
     """ get attribute """
-    def get_edge(self, eid, att=None):
+    def get_edge(self, eid, attrs=None, reset_index=False):
         """Get edge by eid [0, n]"""
-        res = self._get_feature('df_edges', eid, att)
-        if att == 'waypoints':
+        res = self._get_feature('df_edges', eid, attrs, reset_index)
+        if attrs == 'waypoints':
             res = [int(i) for i in res.split(",")]
             
             return res
@@ -61,11 +68,11 @@ class GeoDigraph(Digraph):
         return res
 
 
-    def get_node(self, nid, att=None):
-        return self._get_feature('df_nodes', nid, att)
+    def get_node(self, nid, attrs=None, reset_index=False):
+        return self._get_feature('df_nodes', nid, attrs, reset_index)
 
 
-    def _get_feature(self, df_name, id, att=None):
+    def _get_feature(self, df_name, id, attrs=None, reset_index=False):
         """get edge by id.
 
         Args:
@@ -89,10 +96,37 @@ class GeoDigraph(Digraph):
                     return None
         
         res = df.loc[id]
-        return res if att is None else res[att]
+        
+        if attrs is not None:
+            res = res[attrs]
+        
+        if reset_index:
+            res.reset_index(drop=True, inplace=True)
+        
+        return res
 
 
     """ transfrom """
+    def transform_node_seq_to_edge_seq(self, node_lst:np.array, on:list=['src', 'dst'], key='eid'):
+        if node_lst is None or len(node_lst) <= 1:
+            return None
+        
+        df = pd.DataFrame({on[0]: node_lst[:-1], 
+                           on[1]: node_lst[1:]})
+        
+        # Tips: the merge operation sequence is important
+        eids = df.merge(self.df_edges[on + [key]], on=on)[key].values.tolist()
+
+        return eids
+
+
+    def transform_edge_seq_to_polyline(self, eids):
+        steps = self.get_edge(eids, attrs=['geometry'], reset_index=True)
+        coords = np.concatenate(steps.geometry.apply(lambda x: x.coords), axis=0)
+        
+        return LineString(coords)           
+
+    # deprecated
     def transform_node_seq_to_df_edge(self, node_lst:np.ndarray, on:list=['src', 'dst'], attrs:list=None):
         """Convert the id sequence of nodes into an ordered edges. 
 
@@ -113,7 +147,7 @@ class GeoDigraph(Digraph):
         
         return gpd.GeoDataFrame(df.merge(self.df_edges[on + attrs], on=on))
 
-
+    # deprecated
     def transform_node_seq_to_polyline(self, node_lst:list):
         """Create Linestring by node id sequence.
 
@@ -128,7 +162,7 @@ class GeoDigraph(Digraph):
             return None
         
         steps = self.transform_node_seq_to_df_edge(node_lst, attrs=['geometry'])
-        coords = np.concatenate(steps.geometry.apply(lambda x: x.coords), axis=0)
+        coords = np.concatenate(steps.geometry.apply(lambda x: x.coords).values, axis=0)
         
         return LineString(coords)       
         
@@ -180,16 +214,17 @@ class GeoDigraph(Digraph):
 
     def load_checkpoint(self, ckpt):
         return load_checkpoint(ckpt, self)
-    
+
 
 if __name__ == "__main__":
+    from tilemap import plot_geodata
+    
     network = GeoDigraph()
-    network.load_checkpoint(ckpt='/home/pcl/codes/ST-MapMatching/cache/Shenzhen_graph.ckpt')
+    network.load_checkpoint(ckpt='/home/pcl/codes/ST-MapMatching/cache/Shenzhen_graph_9.ckpt')
 
     route = network.search(src=7959990710, dst=499265789)
-    from tilemap import plot_geodata
-    plot_geodata(network.df_nodes.loc[route['path']])
+    # route = network.search(src=7959602916, dst=7959590857)
 
-    # TODO od 间的 geometry
-    network.transform_node_seq_to_df_edge(route['path'])
-
+    plot_geodata(network.df_edges.loc[route['eid_list']])
+    
+    network.transform_edge_seq_to_polyline(route['eid_list'])